// src/@types/models/article.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conduit API
 * Conduit API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Profile } from './profile';
/**
 * 
 * @export
 * @interface Article
 */
export interface Article {
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    slug: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof Article
     */
    body: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Article
     */
    tagList: Array<string>;
    /**
     * 
     * @type {Date}
     * @memberof Article
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Article
     */
    updatedAt: Date;
    /**
     * 
     * @type {boolean}
     * @memberof Article
     */
    favorited: boolean;
    /**
     * 
     * @type {number}
     * @memberof Article
     */
    favoritesCount: number;
    /**
     * 
     * @type {Profile}
     * @memberof Article
     */
    author: Profile;
}